# Conflicts:
#	jmaqs-webservices-jdk11/pom.xml
#	jmaqs-webservices-jdk11/src/main/java/com/magenic/jmaqs/webservices/jdk11/HttpClientFactory.java
#	jmaqs-webservices-jdk11/src/main/java/com/magenic/jmaqs/webservices/jdk11/WebServiceDriver.java
#	jmaqs-webservices-jdk11/src/main/java/com/magenic/jmaqs/webservices/jdk11/WebServiceUtilities.java
#	jmaqs-webservices-jdk11/src/test/java/com/magenic/jmaqs/webservices/jdk11/WebServiceDriverUnitTest.java
#	jmaqs-webservices-jdk11/src/test/java/com/magenic/jmaqs/webservices/jdk11/models/Product.java

+++ jmaqs-webservices-jdk11/src/test/java/com/magenic/jmaqs/webservices/jdk11/models/Product.java
@@ -7,10 +7,9 @@ package com.magenic.jmaqs.webservices.jdk11.models;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
 
+import java.math.BigDecimal;
+
 public class Product {
-    /**
-     * the xml namespace.
-     */
     @JacksonXmlProperty(isAttribute = true)
     private final String xmlns = "http://schemas.datacontract.org/2004/07/AutomationTestSite.Models";
 
@@ -24,9 +23,9 @@ public class Product {
     private String category;
 
     @JsonProperty("Price")
-    private double price;
+    private BigDecimal price;
 
-    public Product(int id, String name, String category, double price) {
+    public Product(int id, String name, String category, BigDecimal price) {
         this.id = id;
         this.name = name;
         this.category = category;
@@ -59,11 +58,11 @@ public class Product {
         this.category = category;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
